Audit backlog cleanup — F-A5 (vision concurrency), F-A11, F-A12 (#773)
* fix(server): vision request pauses (not stops) the engine — don't cancel concurrent requests

Audit F-A5 (med). A vision request called batching->stop() to get exclusive
engine access while it sets the GLOBAL image, which cancels every concurrent
in-flight request (push_finish("cancelled")). Switch to the pause()/resume()
primitive the embeddings path already uses: pause() drains in-flight generations
and parks the worker (same exclusivity as stop(), verified — the worker only
parks when active_requests_ is empty) WITHOUT cancelling them, and preserves
queued requests.

Every vision exit now clear_image()s before resume() (the image is global;
clearing before resume keeps a resumed/queued request from binding stale vision
embeddings — added clear_image() at the reject paths that previously lacked it).
A pause() timeout returns 503 instead of driving the engine while the worker is
live. The model-reload path keeps stop() (it tears the engine down).

Validated: server boots with gemma-3-4b-vl + mmproj, vision request returns
correct output; a request concurrent with a vision op completes (finish=stop),
not cancelled.


* fix(exec): guard KV-write kernels against negative block_id (defense-in-depth)

Audit F-A12 (low). All 9 paged-KV write kernels resolve block_id via
kv_resolve_slot and index the pool as cache_base + block_id*stride. A negative
block_id (a freed StreamingLLM -1 sentinel, or a future block-table bug) would
write out of bounds. Add a block-uniform `if (block_id < 0) return;` skip after
each resolve. Never fires today (host-side admission guarantees every written
position has a real block, and StreamingLLM -1 slots are skipped host-side), but
converts a silent OOB corruption into a safe no-op — directly hardening the
block-table class of bug just fixed in F-A1.

Validated: GPU suite 0 failures incl StreamingLLM sink eviction + KV-write tests.


* fix(quant): widen calibrate_and_quantize_fp8_async n_elements + guard int overflow

Audit F-A11 (low). The FP8 migrate API took `int n_elements` and callers passed
static_cast<int>(size_t), silently truncating. No current tensor is close
(largest ~778M elems << 2^31), but a >2.1B-element tensor would corrupt with a
wrapped grid + indices. Widen the parameter to int64_t (callers pass the full
size_t) and guard the boundary loudly inside: if n_elements > INT_MAX, log and
skip rather than truncate. The reduction/quantize kernels keep int indexing —
they only ever run for guarded, in-range sizes now.

Validated: build green, FP8 KV oracle tests pass.


* docs(audit): backlog cleanup — F-A5/A11/A12 fixed, pass-1 hygiene already done


---------

+++ src/exec/pre_dequant_phase2_fp8_cache.cu
@@ -195,9 +195,9 @@ void QuantPipeline::pre_dequant_phase2_fp8_cache_(
                 fp8_offset += e.n_elems;
 
                 // Async calibrate + quantize (no host sync)
-                calibrate_and_quantize_fp8_async(qscratch_->dequant, fp8_buf, static_cast<int>(e.n_elems),
-                                                 d_block_maxes, max_grid, d_absmax,
-                                                 d_scales_all + static_cast<ptrdiff_t>(i), stream);
+                calibrate_and_quantize_fp8_async(qscratch_->dequant, fp8_buf,
+                                                 static_cast<int64_t>(e.n_elems), d_block_maxes, max_grid,
+                                                 d_absmax, d_scales_all + static_cast<ptrdiff_t>(i), stream);
 
                 int64_t fp8_shape[4] = {e.weight.shape[0],
                     (e.qtype == QType::NVFP4) ? e.weight.shape[1] * 2 : e.weight.shape[1],
